requirements.txt
cmd2func>=0.1.4
oneface>=0.2.2
executor-http>=0.1.0
fire  
rich
jinja2
pyYAML
